refactor: extract CLI code from main into run to compile fake for testing purposes
The code in `main` is moved into `run`, which now exposes two versions of
run_cli, one enabled in production and the other one only for testing purposes,
using a fake implementation of the `api` trait
