+++ iroh-ctl/src/lib.rs
@@ -1,6 +1,9 @@
 pub mod config;
+#[cfg(test)]
+mod fake;
 pub mod gateway;
 pub mod metrics;
 pub mod p2p;
+pub mod run;
 pub mod status;
 pub mod store;
